Refactor params validations
