--- app/validators/schemas/tasks/list_of_tasks_schema.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Schemas
-  module Tasks
-    ListOfTasksSchema = Dry::Schema.Params do
-      required(:to_do_list_id).filled(:string)
-    end
-  end
-end
