--- app/validators/schemas/tasks/single_task_schema.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module Schemas
-  module Tasks
-    SingleTaskSchema = Dry::Schema.Params do
-      required(:to_do_list_id).filled(:string)
-      required(:id).filled(:string)
-    end
-  end
-end
