--- app/validators/schemas/tasks/update_task_schema.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Schemas
-  module Tasks
-    UpdateTaskSchema = Dry::Schema.Params do
-      required(:to_do_list_id).filled(:string)
-      required(:id).filled(:string)
-      optional(:title).filled(:string)
-    end
-  end
-end
