+++ app/validators/tasks/delete_params_validator.rb
@@ -1,11 +1,10 @@
 # frozen_string_literal: true
 
 module Tasks
-  class DeleteParamsValidator < Common::BaseValidator
-    params(Schemas::Tasks::DeleteTaskSchema)
-
-    register_macro(:uuid_format) do
-      key.failure(:invalid_format, field: :id, format_name: :UUID) unless UUID_FORMAT.match?(value)
+  class DeleteParamsValidator < BaseValidator
+    params do
+      required(:to_do_list_id).filled(:string)
+      required(:id).filled(:string)
     end
 
     rule(:to_do_list_id).validate(:uuid_format)
