+++ app/validators/tasks/update_params_validator.rb
@@ -1,17 +1,11 @@
 # frozen_string_literal: true
 
 module Tasks
-  class UpdateParamsValidator < Common::BaseValidator
-    params(Schemas::Tasks::UpdateTaskSchema)
-
-    register_macro(:uuid_format) do
-      key.failure(:invalid_format, field: :id, format_name: :UUID) unless UUID_FORMAT.match?(value)
-    end
-
-    register_macro(:max_length) do |macro:|
-      max = macro.args[0]
-
-      key.failure(:exceeds_maximum_length, field: key.path.keys[0], length: max) if value && value.size > max
+  class UpdateParamsValidator < BaseValidator
+    params do
+      required(:to_do_list_id).filled(:string)
+      required(:id).filled(:string)
+      optional(:title).filled(:string)
     end
 
     rule(:to_do_list_id).validate(:uuid_format)
