+++ app/validators/to_do_lists/show_params_validator.rb
@@ -1,12 +1,8 @@
 # frozen_string_literal: true
 
 module ToDoLists
-  class ShowParamsValidator < Common::BaseValidator
+  class ShowParamsValidator < BaseValidator
     params(Common::Schemas::OnlyIdParamSchema)
-
-    register_macro(:uuid_format) do
-      key.failure(:invalid_format, field: :id, format_name: :UUID) unless UUID_FORMAT.match?(value)
-    end
 
     rule(:id).validate(:uuid_format)
   end
